python_backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy-serializer==1.4.22
Flask-Session==0.8.0
PyJWT==2.10.1
python-dotenv==1.1.0
sendgrid==6.12.0
python-dateutil==2.9.0
Werkzeug==3.1.3